+++ src/sensai/evaluation/evaluator.py
@@ -1,7 +1,7 @@
 import functools
 import logging
 from abc import ABC, abstractmethod
-from typing import Tuple, Dict, Any, Generator, Generic, TypeVar, Sequence, Optional, List, Union, Callable
+from typing import Tuple, Dict, Any, Generator, Generic, TypeVar, Sequence, Optional, List, Union, Callable, Hashable
 
 import numpy as np
 import pandas as pd
@@ -506,7 +506,7 @@ class VectorClassificationModelEvaluationData(VectorModelEvaluationData[Classifi
 class ClassificationEvaluatorParams(EvaluatorParams):
     def __init__(self, data_splitter: DataSplitter = None, fractional_split_test_fraction: float = None, fractional_split_random_seed=42,
             fractional_split_shuffle=True, additional_metrics: Sequence[ClassificationMetric] = None,
-            compute_probabilities: bool = False, binary_positive_label: Optional[str] = GUESS):
+            compute_probabilities: bool = False, binary_positive_label: Optional[Hashable] = GUESS):
         """
         :param data_splitter: [if test data must be obtained via split] a splitter to use in order to obtain; if None, must specify
             fractionalSplitTestFraction for fractional split (default)